add make dataset slug to readable name filter factory function and move into separate file for easier testing
